test: Deal with /bin, /sbin unification in Fedora
Fedora is apparently merging /bin and /sbin together [0].  At least some of
this has landed in Fedora 42, with sshd and dhclient now being in /bin.
They are symlinked from /sbin, but mbuto doesn't preserve that symlink when
it builds its image, it just puts them in /bin.  That breaks later parts
of the script which expects to find sshd and dhclient in /sbin.  Other
parts of the test also expect sysctl in /sbin.

Use mbuto's LINKS feature to deal with these cases.

[0] https://fedoraproject.org/wiki/Changes/Unify_bin_and_sbin

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>