test: Explicit specify forwarding ports for pasta in log rotation tests
test/pasta_options/log_to_file uses tcp_crr in a loop to generate log
messages for what it's actually testing.  For this to work, pasta needs to
forward two ports.  We set a PORTS variable with the options for this, but
never actually use it.  The test works, because without the options we use
automatic port forwarding.

That makes the test more fragile than it needs to be, in a way that's not
at all related to what we're actually trying to test.  Avoid this possible
point of failure by explicitly specifying the forwarding.

I encountered this because Fedora has now updated to a kernel with the
/proc/net lseek() bug.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>