Add missing explicit PSH assignment
For packets with data, we set the PSH flag depending on whether it's
the last chunk of data in a single receive operation from a socket.
But then we reuse the same buffers for SYN segments, and oops, PSH is
set randomly.

This change cleans this up by setting PSH explicitly, just as we do
with the other flags.

Signed-off-by: Volker Diels-Grabsch <v@njh.eu>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>