test: Update lib/term for clearer output when DEBUG is enabled
When running tests with DEBUG=1, we used to get some
"DEBUG:DEBUG:DEBUG:" in logs, which is not very helpful for
debugging.

With this patch, the log would show the steps like below:

    DEBUG: g1tools  ip jq dhclient socat cat
    DEBUG: htools   ip jq
    DEBUG:
    DEBUG: set      MAP_HOST4 192.0.2.1
    DEBUG: set      MAP_HOST6 2001:db8:9a55::1

The problem here is that "${@}" is magic and expands to multiple
arguments, even though it's quoted.  That means instead of the whole
message going into a single string, the "DEBUG:" becomes the format
string and the rest goes into unused parameters.

Suggested-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Yumei Huang <yuhuang@redhat.com>
[dwg: root cause explanation in commit message]
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>