test: Update mbuto profile to fix the symlink of /bin
When running test 'migrate/rampstream_out' on fedora, test fails
with error like this:

  guest_1$ socat -u EXEC:"rampstream send 6000000" TCP4:192.0.2.1:10006
  2025/09/24 04:06:12 socat[405] E execvp("rampstream", "rampstream", "send", "6000000", ...): No such file or directory

It's because on fedora, /bin is not in the PATH, and some binaries
are under /usr/bin. mbuto will find the binaries on the host and put
them into the guest under same directories. As a result, /usr/bin in
mbuto guest are not empty, causing creating /usr/bin symlink failed.
Thus, the script 'rampstream' under /bin is not found during testing.

This patch fixes the issue and creates the symlink for consistency
with /sbin.

Link: https://bugs.passt.top/show_bug.cgi?id=151
Suggested-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Yumei Huang <yuhuang@redhat.com>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>