test: Update passt.mbuto and passt.mem.mbuto
If /bin is in the PATH of host, /usr/bin may not exist in mbuto
guest, which causes some binaries failing to move to /usr/bin.
As a result, tests fail as binaries not found.

Fix it by adding /usr/bin to DIRS. Also create the same symlink
for mbuto.mem.img.

Link: https://bugs.passt.top/show_bug.cgi?id=158
Fixes: f4729be7a56b ("test: Update mbuto profile to fix the symlink of /bin")
Suggested-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Yumei Huang <yuhuang@redhat.com>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>